Move SocialLinkProps to SocialLink file
