+++ src/sections/home/Landing.tsx
@@ -19,8 +19,7 @@ import { Heading, Flex, Box, Text } from 'rebass/styled-components';
 
 import { Fade } from 'react-awesome-reveal';
 
-import { Section } from '../../theme/components/Section';
-import { SocialLink } from '../../theme/components/SocialLink';
+import { Section, SocialLink } from '../../theme/components';
 import { SECTION } from '../../theme/utils/constants';
 
 import { landing } from '../../content/LandingContent';
